agent_registry_service/requirements.txt
flask==3.0.0
weaviate-client==4.9.3
requests>=2.31.0
python-dotenv==1.0.0
